[Midway] Shift column inference logic for CSV import to CSV parsing step instead of saving CSV content step
